+++ src/collider/pointCollider/pointCollider.h
@@ -6,11 +6,11 @@ class PointCollider : public Collider
 public:
 	PointCollider(Surface* pScreen, LevelMaps* pLevelMaps) : Collider(pScreen, pLevelMaps) {}
 
-	void UpdatePosition(int2 pos);
+	void UpdatePosition(float2 pos);
 	virtual void Draw(int size, int color = 0xff0000);
-	int2 GetPos() const { return m_pos; }
+	float2 GetPos() const { return m_pos; }
 	TileType GetTileType() const { return m_pLevelMaps->GetTileType(m_pos); }
 
 private:
-	int2 m_pos;
+	float2 m_pos;
 };
